+++ release-notes.md
@@ -4,6 +4,7 @@ RELEASE NOTES
 - update webdriver version 3.3.1
 - add FieldSet for extjs 6 package
 - add check(Cell... cells) and unCheck(Cell... cells) in Grid and Table
+- check(Cell... cells) and unCheck(Cell... cells) doing scrollButton if row is not visible in Grid
 
 **Release Notes for Testy 2.3.0**
 - update webdriver version 3.0.1