+++ release-notes.md
@@ -1,11 +1,18 @@
 RELEASE NOTES
 -------------
-**Release Notes for Testy 2.4.0-SNAPSHOT**
+**Release Notes for Testy 2.5.0-SNAPSHOT**
+- update webdriver version 3.4.0
+
+**Release Notes for Testy 2.4.0**
 - update webdriver version 3.3.1
 - add FieldSet for extjs 6 package
 - add check(Cell... cells) and unCheck(Cell... cells) in Grid and Table
 - check(Cell... cells) and unCheck(Cell... cells) doing scrollButton if row is not visible in Grid
+- added method scrollToWebLocator(WebLocator element) in Utils
 - added  method getError for TextField in extjs 6 package
+- remove deprecated methods
+- added select(Row row) and unSelect(Row row) in Grid
+- added Cell(String columnText, SearchType... searchTypes) method in Cell
 
 **Release Notes for Testy 2.3.0**
 - update webdriver version 3.0.1